Made a number of fixes to VVER
Firstly, the gap_cell wasn't filled with any material. It needs to be filled with helium.

Secondly, the assembly and reflector hexagonal prism were the same size, so the reflector had zero width. I made it 1 cm. If you don't want a reflector, that's fine. You can delete it, but then you will need boundary conditions on the assembly.

Thirdly, the moderator shouldn't be confined to a circle around the fuel, because you will also want it to fill all the areas of the lattice between the fuel rods.

Fourthly, you can't reuse fuel region, sorry! It's a pain, but you have to redefine it each time so that the same surface doesn't get reused in each universe.

I made these changes for you. Let me know if you have questions